+++ server/src/routes/index.ts
@@ -32,3 +32,4 @@ export { accessRoutes } from "./access.js";
 export { instanceSettingsRoutes } from "./instance-settings.js";
 export { instanceDatabaseBackupRoutes } from "./instance-database-backups.js";
 export { cloudUpstreamRoutes } from "./cloud-upstreams.js";
+export { runBearerRoutes } from "./run-bearer.js";